fix bug when pasting items (which contain completed items) into a collapsed item parent
we have to create items, instead of duplicating them, when pasting them into a collapsed parent. The process of creating items recursively has certain requirements. In this case the requirement we ran into was that the created item, which we wanted to mark as completed didn't have a expanded parent. Obviously the solution is to expand the parent - but expanding items is coupled with a transition animation. Which is unfortunate because that causes ~0.1ms lag every time 1 item is created. We either want to cancel the animation (but we need the handle for that right?) or wait for workflowy to provide us with a expandItem() which skips the animation part. For now we'll have to live with the lag