WIP instant folding
We got it working, but it wasn't as visually pleasing as expected.

There was no performance gain from killing the expansion animation/transition/transformation.

WF.expand/collapse still trigger transition timeouts - even though the CSS tells it not to, and it doesn't visually animate neither...

stopping the expansion/collapse timeouts gave us wierd bugs. A potential solution is to override the window.setTimeout() function, enabling us to cache the function that was supposed to be called upon clearing the timeout. But that is a lot of work. But it is the next stepll