add a naive 'daw' binding + increment version to 0.0.2.0
we changed the mode handling to 'inner' and 'after' instead. Next time we'll rewrite everything to be handled via a single 'modify text object' state, where all substate need to go through. Thats the most optimized solution so far and it gives us a better overview by branching the text manipulation to it own hierarchical branch