make moveSelectionDown almost as fast as moveSelectionUp
moving a selection down can result in focus loss because the next item might be an item with a lot of kids, that stretch beyond the screen. Once we move off screen we lose focus. We do a preemptive check, to ensure that we don't error out, by moving 2 steps instead of 1 and checking if we are still in focus. Now, if we lose focus then we move back and collapse the item in front of us. This whole operation is quite expensive -- and it was being done every time EVEN if the next item didn't have any kids! So we don't do that for items that don't have any kids anymorell