Promote the use of 'ctrl-c' to copy items between browsers and non original items, such as mirrors
We do not save the selected items to the 'yankBuffer' when this happens, and vice versa. It was tempting, but it will cause problems. One can easily alter the code to enable it though, if prefered.

yanking and pasting mirrors results in pasting copies of their original item. Sometimes one might prefer to copy-paste an actual mirror. To support that in the 'pipeline' would be tedious. A much simpler solution was to promote the use of 'ctrl-c' when one wants to copy-paste mirrors

'ctrl-c', while in normal mode, will copy the entire item rather then the selected text. Copy-pasting text still works because you'll enter insert mode immediately when using the mouse on the webpage, so you won't be normal mode anymore when that happens.