bugfix: creating an item below the (expanded) focusedItem, by pressing 'o', will now work
the previous behaviour was that it would create a child within the expanded item instead... I guess that felt convenient at the time. Problem is that it doesn't mirror the 'O' behaviour. I'd rather not have inconsistencies.

There is a problem with the new behaviour though; it will fail if the expanded focused item's kids strech beyond the screen. Quick fix for this is to collapse the focused item and have the user press twice. We tried scrolling downwards to get past child items - but you can only scroll to the bottommost visible item and you only do it once